Release v0.4.4: native LuaTeX engine, automatic convergence, and in-tree PostScript/MetaPost

+++ crates/tex-core/src/boxes.rs
@@ -449,7 +449,7 @@ pub enum Node {
     },
     MathChar {
         fam: u8,
-        c: u8,
+        c: u32,
         class: u8,
         origin: MathDiagnosticOrigin,
     },
